LayoutTests/http/tests/security/contentSecurityPolicy/register-bypassing-scheme.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="X-WebKit-CSP" content="img-src https:; script-src 'unsafe-inline'">
<script>
    if (window.testRunner) {
        testRunner.waitUntilDone();
        testRunner.dumpAsText();
    }

    function test1() {
        var img = document.createElement('img');
        img.onload = function () { alert('FAIL'); };
        img.onerror = function () { alert('PASS (1/3)'); };
        try {
            img.src = "../resources/abe.png";
        } catch (e) {
            alert("PASS (1/3)");
        }
        test2();
    }

    function test2() {
        internals.registerURLSchemeAsBypassingContentSecurityPolicy('http');
        var img = document.createElement('img');
        img.onload = function () {
            alert('PASS (2/3)');
            test3();
        };
        try {
            img.src = "../resources/abe.png";
        } catch (e) {
            alert('FAIL');
            test3();
        }
    }

    function test3() {
        internals.removeURLSchemeRegisteredAsBypassingContentSecurityPolicy('http');
        var img = document.createElement('img');
        img.onload = function () { alert('FAIL'); };
        img.onerror = function () { alert('PASS (2/3)'); };
        try {
            img.src = "../resources/abe.png";
        } catch (e) {
            alert("PASS (3/3)");
        }
        finishTesting();
    }

    function finishTesting() {
        if (window.testRunner) {
            setTimeout(function () { testRunner.notifyDone(); }, 0);
        }
        return true;
    }
</script>
</head>
<body onload='test1();'>
    <p>
        This test ensures that registering a scheme as bypassing CSP actually bypasses CSP.
        This test passes if three PASSes are generated.
    </p>
</body>
</html>
