Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/register-bypassing-scheme-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/resources/abe.png because it does not appear in the img-src directive of the Content Security Policy.
ALERT: PASS (1/3)
ALERT: PASS (2/3)
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/resources/abe.png because it does not appear in the img-src directive of the Content Security Policy.
ALERT: PASS (3/3)
This test ensures that registering a scheme as bypassing CSP actually bypasses CSP. This test passes if three PASSes are generated.
